pyproject.toml
[project]
name = "svansokning"
version = "0.1.0"
description = "Personal Oxford rental dashboard"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.40",
    "httpx>=0.27",
    "selectolax>=0.3.21",
    "pandas>=2.2",
    "pyarrow>=16",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/svansokning"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
